requirements.txt
httpx>=0.27.0
websockets>=12.0
python-dotenv>=1.0
